+++ app/services/create_review_notification_service.rb
@@ -1,0 +1,9 @@
+module CreateReviewNotificationService
+  def create_notification_review!(current_user, user)
+    notification = current_user.active_notifications.new(
+      visited_id: user.id,
+      action: 'review'
+    )
+    notification.save if notification.valid?
+  end
+end
